Don't include stopped apps and remove any apps that stop
Includes a `Consistently` test condition to ensure that we do not try and
delete an appWatcher twice. This test condition was put in place
because we ran into a bug where after an app was stopped it's app
watcher was closed but then on the next run of `checkForNewApps`
it would try and delete the app watcher a second time. This was
because we have two places where app watchers are tracked. We
mitigated this by explicitely removing it from the `appNameByGuid`
map as well as by calling `DeleteWatcher`. We suspect this bug
also would have appeared when deleting apps entirely rather than
just stopping them.