Ensure the state of which apps are to be watched is correct
As there are two places of truth for this (appWatchers and
`appNameByGuid` map), we create a new function to handle
deletion of an app watcher that ensures the state is removed
from both of those sources.

Also adds an additional test to check that a stopped app that
is started gets watched.