Add crashes counter
Follows very similar logic to our other instance metrics. The
only difference is the source of data is from log message events
rather than container metric events. To support this we have
copied and adapted most of the `log_message_processor` and its
tests from the paas-metric-exporter (see
https://github.com/alphagov/paas-metric-exporter/blob/master/processors/log_message_processor.go).

The crash counter increments by 1 each time it picks up a crash log, we do this rather than
using crash_count from the log because we want to keep track of total crashes of an app even
if a redeploy happens.