Add requests and response time metrics
This adds in requests and response time metrics broken down into status ranges 2, 3, 4 and 5xx. They are all initialised with a zero value.  The paas-metric-exporter had functionality for 1xx and other response codes. We have elected not to do this because we never expect to have these codes.

In our tests we moved the creation of the message chan into the before each to avoid race conditions.

We have not tested response times explicitly, this is because it was not possible to get private data about buckets and their values from the response_time histogram. We looked into using the registerer but were not able to do a more integration style test using this method because it's being mocked. We decided the time needed to have good tests for this was not worth the benefit. Our changes have been manually tested and we get all the information we were expecting.

https://trello.com/c/wy3K0DB7/741-paas-prometheus-exporter-requests-and-response-times