error out if the noaa channels are closed
If these channels are closed, it's because there is a non-recoverable
error in the doppler stream client.  At this point there's nothing
much we can do other than propagate the error back up and have it
crash the app with log.Fatal().

There were two problems with the previous behaviour:

1. possibility of double-closing a channel

Calling m.Close() when the messages channel is closed was a cute way
of getting us to scale down to 0 metrics.  However, it causes a
hazard, because the consumer (ie the Exporter) doesn't know that the
appInstancesChannel has been closed and so may well try to close it
again in future.  This would be a second close on an already-closed
channel, which causes a panic.

2. ignoring unrecoverable problems

When the noaa channels (ie, msgs and errs) are closed, this is a sign
that noaa has encountered a problem it can't deal with - eg, the
websocket connection has failed.  If we just call `continue` and go
round the loop again, we leave the AppWatcher in a state where it
won't get any more metrics, but the Exporter doesn't know that there
have been any problems.

The fix is simple: when the channels are closed, we exit the
AppWatcher mainLoop() with an error which will be logged out with
log.Fatal() in AppWatcher.Run(), crashing the whole app so that we can
do a clean restart.