Change HTTP client used for log-cache
log-cache was using the same HTTP client as the app metrics client - this was causing the Authorization header we set to be overwritten with 'Bearer $token'.

Apparently log-cache no longer likes `Bearer` to be present in the header value.

Using `http.DefaultClient` rather than the one cfclient generates resolves this.