Ignore HttpStartStop messages with PeerType = Server
We always receive two HttpStartStop messages per requests. One has a peer type
of client, and the other has a peer type of server. The former includes an
instance index, indicating which application index the request was routed to.
The latter does not, but the library we use converts the nil to a zero.

This was causing every HttpStartStop message with PeerType = Server to be
interpreted as a request routed to app instance zero. This resulted in the
HTTP requests metrics being reported as N+1, where N is the number of requests
across all app instances.