Add retries when getting an oauth token
From time to time, the exporter will fail to make a connection to UAA for some
transient reason. In most cases, the next attempt would be fine, as judged by
the logs. This commit wraps some retry logic around getting the token to
improve the reliability of discovery of apps and services by the exporter.